pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "自然语言控制 PyCharm IDE 系统"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyautogui>=0.9.54",
    "keyboard>=0.13.5",
    "pillow>=10.0.0",
    "mss>=9.0.1",
    "zhipuai>=2.0.0",
    "pyyaml>=6.0.0",
    "pydantic>=2.0.0",
    "easyocr>=1.7.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "easyocr>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
